HUN_ADM2 #1839
HUN_ADM2 #1839

Budapest is called the capital city and not a county so I'm not totally sure how to address that. It's surrounded by the county of Pest but I'm not sure if Pest includes Budapest? I think they are administered separately so I added Budapest as its own entry. Just a heads up!